src/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
pydantic
pydantic-settings
chromadb
qdrant-client
groq
openai
langchain-core
langchain-community
langchain-groq
langchain-openai
sqlite3
uuid
chromadb
